add editinterface

+++ src/com/buaa/hospital/nomogram/PredictButtonListener.java
@@ -47,7 +47,7 @@ public class PredictButtonListener implements ActionListener{
 				}
 			}
 			attribute.setPredictNomogram(Nomogram);
-			predictInterface.setLogText("病历号为 " + attribute.getID() + " 的病人 " + attribute.getName() + " 使用  " + predictInterface.getAlogrithmName() + " 的预测球镜调整值为 " + Nomogram);
+			predictInterface.setLogText("病历号为 " + (long)attribute.getID() + " 的病人 " + attribute.getName() + " 使用  " + predictInterface.getAlogrithmName() + " 的预测球镜调整值为 " + Nomogram);
 			predictInterface.addHistory(attribute);
 			predictInterface.RefreshHistoryTable();
 		}
